refactor(settings): move the audiosettings to containers
